refactor(POST): refactor on post for positions
